move run-time config into config folder
