+++ src/app.module.ts
@@ -42,7 +42,7 @@ import {
 } from "./common/schemas/generic-history.schema";
 import { HistoryModule } from "./history/history.module";
 import { MaskSensitiveDataInterceptorModule } from "./common/interceptors/mask-sensitive-data.interceptor";
-import { RuntimeConfigModule } from "./runtime-config/runtime-config.module";
+import { RuntimeConfigModule } from "./config/runtime-config/runtime-config.module";
 
 @Module({
   imports: [
